EJ_SELECT_VARIABLES_C2000_ACS2014_COMPARE_DEC28
Script Purpose: To create a table which provides census variables for 2000 and 2010, in
terms of 2010 tract boundaries. Data provided within these variables can be used to understand
changes between 2000 and 2010. 

This script seeks to surface logic used to apportion 2000 census figures to 2010 census tracts
which may or may not have been split, changed, or motified between 2000 and 2010. 

**Script Adapted from 'ACS Census 2010 to Census 2000 Corespondence.sql'**
Created On: 12/30/2016
Created By: Josh Croff